ai-analysis-service/requirements.txt
# Core
opencv-python-headless==4.9.0.80
mediapipe==0.10.11
numpy==1.26.4

# Kafka
kafka-python==2.0.2

# HTTP client
httpx==0.27.0

# API server (health endpoint)
fastapi==0.110.1
uvicorn[standard]==0.29.0

# Azure Blob Storage
azure-storage-blob==12.19.1

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
loguru==0.7.2
